Fix ReporteRepository id type

+++ ms-reportes/src/main/java/cl/hilton/reportes/repository/ReporteRepository.java
@@ -1,15 +1,15 @@
 package cl.hilton.reportes.repository;
 
-
-import cl.hilton.reportes.model.Reporte;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import cl.hilton.reportes.model.Reporte;
+
 @Repository
-public interface ReporteRepository extends JpaRepository<Reporte, Integer> {
+public interface ReporteRepository extends JpaRepository<Reporte, Long> {
 
     Optional<Reporte> findByCodigo(String codigo);
